+++ node-net-manager/table_entry_cache/translation_table_test.go
@@ -1,4 +1,4 @@
-package TableEntryCache
+package table_entry_cache
 
 import (
 	"fmt"
